Increase diameter of motor cross pad
I am increasing the diameter of the motor cross pad by a few mm.
I am doing this because I have had two crashes where two different
mounts have broken in strikingly similar ways. My theory is that
the pad is flexing and one of the screw heads is making contact with
the cutout wall. When this happens the screw doesn't flex and causes
the mount to crack.