Mark 6 improvements
There were several improvements in the iteration

Most obvious is that I added a landing strut to the bottom of the
motor mount. This is in preperation for a full redesign of the
quadcopter that will move the landing gear out to the ends of the
arms.

Next I added a center hole to the motor cross. My previous motors
didn't have any of there center spindle protruding however I
picked up some different motors recently that do so I need a hole
in the center so that the spindle doesn't rub.

One other minor update is that I made changes to the pvc adapter
module so now it can have a setable number of flat sides. I need
to update the arguments to reflect this change.

Additionally I am increasing the size of the side mount holes to
accomidate new self locking trailer pins I want to try instead of
screws.