Fix issue with hex construction height
In trying to make the pocket for the nut deeper on the bracket I
found that small changes weren't having the desired result upon
looking at the model I discovered that the reason was that half
the length of the hexagon is below the z plane to start with
meaning every dimension change I made was being cut in half. I
am fixing this in the hex function by translating the part up
by half the height before exiting the function.

I also found I wasn't unioning the different boxes so they were
just being group in my models by accident.