Add max removal base on right tangent trapizoid
Currently the circles in the web of the strut are just created
based on a simple algorithm that determines the max size base
on some other parameters. I am converting this method to one
that uses the properties of a right tangential trapizoid in order
to find the size of each trapizoid so that a circle that is tangent
to all sides can be created maximizing the amount of material
that can be removed.

Using elipses might yield even more material remove but this
solution is pretty optimal.